requirements.txt
python-dotenv
selenium
requests
beautifulsoup4